requirements.txt
blinker==1.9.0
click==8.1.8
flask==3.1.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
werkzeug==3.1.3
Pillow
